requirements.txt
playwright
behave
allure-behave
allure-python-commons
cucumber_tag_expressions
pytest-playwright
allure-pytest
